Bugfix: allow copying contents when Quill is disabled
Fixes: https://github.com/voxmedia/anthem/issues/3534

Our custom copy logic relies on a quill `range` to retrieve the
plaintext and JSON data to be stored in the clipboard. In turn,
Selection.getRange, the method that ultimately provides this range,
returns `null` when quill is disabled. This is because users cannot
focus on the quill instance root when disabled.

To resolve this issue, I've added a new option to be passed down to
the getRange method so that we always try to retrieve a selection
range, even if the user isn't technically focused on @doc.root.